requirements.txt
streamlit
gooogle-generativeai
pandas
python-dotenv

